Added generics to the entire genometryImpl module.
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@2853 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ genometryImpl/src/com/affymetrix/genometryImpl/parsers/gchp/AffyCnChpParser.java
@@ -87,7 +87,7 @@ public class AffyCnChpParser {
             for (int i=0; i<positions.size(); i++) {
               //TODO: insn't there a class that accepts an IntList as the set of positions?
               final int start = positions.get(i);
-              SingletonSymWithProps sym = new SingletonSymWithProps(data.probeSetNames.get(i), (int) start, (int) start+1, seq);
+              SingletonSymWithProps sym = new SingletonSymWithProps(data.probeSetNames.get(i), start, start+1, seq);
               sym.setProperty("method", theMethod);
 //          tca.addChild(sym);
               seq.addAnnotation(sym);
